PageSprite, PageScene 等维护图片变换信息的类中，所有像素单位和比例单位均改为 qreal

+++ src/gui/controller/controller.h
@@ -108,9 +108,9 @@ signals:
 
     void cmdTranslateBy(qreal dx, qreal dy);
 
-    void sigScaleRangeUpdated(float min, float max);
+    void sigScaleRangeUpdated(qreal min, qreal max);
 
-    void sigScaleUpdated(float scale);
+    void sigScaleUpdated(qreal scale);
 
     void guiEngineChanged();
